Unit Debug: Remove need for traits_intial
This also changes behavior, as the unit traits will always be removed and readded.
To accomadate this, and work according to expectations, the restoration of hp and moves on trait change is removed.

The loss of this functionality will be returned in a future feature.
Also, bug where not changing anything can cause loss of hp needs to be fixed.
This bug is due to new traits being applied one at at time, and if quick is first, it lowers max hp and hp with it.